requirements.txt
virtualenv
django
pillow
django-auditlog